Socorro

+++ Controllers/UsuarioController.cs
@@ -3,6 +3,7 @@ using aspnet_store.Models.Entities;
 using aspnet_store.Models.ViewModels.UsuarioViewModels;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using System.Threading.Tasks;
 
 namespace aspnet_store.Controllers
 {
@@ -12,13 +13,18 @@ namespace aspnet_store.Controllers
 
         public UsuarioController(ApplicationDbContext dbContext) => this.dbContext = dbContext;
 
+        public IActionResult Index()
+        {
+            return View();
+        }
 
         [HttpGet]
         public IActionResult Add()
         {
+            var departamentos = dbContext.Departamentos.ToList();
             var viewModel = new AddUsuarioViewModel
             {
-                Departamentos = dbContext.Departamentos.ToList()
+                Departamentos = departamentos
             };
 
             return View(viewModel);
@@ -28,6 +34,7 @@ namespace aspnet_store.Controllers
         public async Task<IActionResult> Add(AddUsuarioViewModel viewModel)
         {
             ModelState.Remove(nameof(viewModel.Departamentos));
+            ModelState.Remove(nameof(viewModel.DepartamentoNome));
             if (ModelState.IsValid)
             {
                 var usuario = new Usuario
@@ -43,7 +50,6 @@ namespace aspnet_store.Controllers
                 return RedirectToAction("List", "Usuario");
             }
 
-            viewModel.Departamentos = dbContext.Departamentos.ToList();
             return View(viewModel);
         }
 
@@ -51,7 +57,6 @@ namespace aspnet_store.Controllers
         public async Task<IActionResult> List()
         {
             var usuarios = await dbContext.Usuarios.Include(u => u.Departamento).ToListAsync();
-
             return View(usuarios);
         }
 
@@ -61,7 +66,9 @@ namespace aspnet_store.Controllers
             if (id is null)
                 return NotFound();
 
-            var usuario = await dbContext.Usuarios.FindAsync(id);
+            var usuario = await dbContext.Usuarios
+                .Include(u => u.Departamento)
+                .FirstOrDefaultAsync(u => u.Id == id);
 
             if (usuario is null)
                 return NotFound();
@@ -69,39 +76,59 @@ namespace aspnet_store.Controllers
             var viewModel = new AddUsuarioViewModel
             {
                 Id = usuario.Id,
-                Matricula = usuario.Matricula,
                 Nome = usuario.Nome,
+                Matricula = usuario.Matricula,
                 DepartamentoId = usuario.DepartamentoId,
-                Departamentos = dbContext.Departamentos.ToList()
+                Departamentos = await dbContext.Departamentos.ToListAsync()
             };
 
             return View(viewModel);
         }
 
         [HttpPost]
-        public async Task<IActionResult> Edit(Usuario viewModel)
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> Edit(AddUsuarioViewModel viewModel)
         {
-            var usuario = await dbContext.Usuarios.FindAsync(viewModel.Id);
-
-            if (usuario is not null)
+            ModelState.Remove(nameof(viewModel.Departamentos));
+            ModelState.Remove(nameof(viewModel.DepartamentoNome));
+            if (ModelState.IsValid)
             {
-                usuario.Nome = viewModel.Nome;
-                usuario.Matricula = viewModel.Matricula;
-                usuario.DepartamentoId = viewModel.DepartamentoId;
+                var usuario = await dbContext.Usuarios.FindAsync(viewModel.Id);
 
-                await dbContext.SaveChangesAsync();
+                if (usuario is not null)
+                {
+                    usuario.Nome = viewModel.Nome;
+                    usuario.Matricula = viewModel.Matricula;
+                    usuario.DepartamentoId = viewModel.DepartamentoId;
+
+                    await dbContext.SaveChangesAsync();
+                    return RedirectToAction("List", "Usuario");
+                }
             }
 
-            return RedirectToAction("List", "Usuario");
+            viewModel.Departamentos = await dbContext.Departamentos.ToListAsync();
+            return View(viewModel);
         }
 
+
+
+        [HttpPost]
+        [ValidateAntiForgeryToken]
         public async Task<IActionResult> Delete(int id)
         {
-            var usuario = await dbContext.Usuarios
-                .FirstOrDefaultAsync(i => i.Id.Equals(id));
+            var usuario = await dbContext.Usuarios.FindAsync(id);
 
             if (usuario is not null)
             {
+                var pedidosAssociados = await dbContext.Pedidos
+                    .AnyAsync(p => p.UsuarioSolicitanteId == id);
+
+                if (pedidosAssociados)
+                {
+                    TempData["ErrorMessage"] = $"O usuário {usuario.Nome} tem pedidos associados. Não pode ser excluído.";
+                    return RedirectToAction("List");
+                }
+
                 dbContext.Usuarios.Remove(usuario);
                 await dbContext.SaveChangesAsync();
             }
